tests/wpt/tests/html/semantics/interactive-elements/the-dialog-element/dialog-closedby-start-open.html

<!doctype html>
<meta charset="utf-8">
<meta name="timeout" content="long">
<link rel=help href="https://html.spec.whatwg.org/multipage/interactive-elements.html#dialog-light-dismiss">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/resources/testdriver.js"></script>
<script src="/resources/testdriver-actions.js"></script>
<script src="/resources/testdriver-vendor.js"></script>
<script src="../../popovers/resources/popover-utils.js"></script>

<dialog id=test1 open closedby="any"></dialog>

<script>
const ESC = '\uE00C';
promise_test(async (t) => {
  const dialog = document.querySelector('dialog#test1');
  assert_true(dialog.open);
  await new test_driver.send_keys(document.documentElement,ESC);
  assert_false(dialog.open);
  dialog.showModal();
  assert_true(dialog.open);
  await new test_driver.send_keys(document.documentElement,ESC);
  assert_false(dialog.open);
}, `Dialogs that start open and have closedby should still function`);
</script>

<dl>
  <dt contenteditable></dt>
  <dialog id=test2 open></dialog>
</dl>

<script>
promise_test(async (t) => {
  // This test case is pulled from `dialog-closewatcher-crash.html`.
  const dialog = document.querySelector('dialog#test2');
  document.querySelector('dl').addEventListener("focusin", () => {
    dialog.showModal();
  });
  // This will trigger the focus-the-previous-element behavior, which will fire
  // the `focusin` event.
  dialog.open = false;
  await new Promise(resolve => {
    document.defaultView.requestIdleCallback(() => {
      window.getSelection().addRange(document.createRange());
      dialog.close();
      resolve();
    });
  });
  assert_true(dialog.open);
  await new test_driver.send_keys(document.documentElement,ESC);
  assert_false(dialog.open,'ESC should still work');
}, `Opening and closing a dialog during the dialog focus fixup should still leave closedby functional`);
</script>
